added the entire styles
I've implemented the design for all pages that needed to be designed. I encountered some problems with some of the Material UI elements that messed up my entire project and I decided to refactor it and styling it using SCSS, because of fear of losing too much time on this.